email/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
